AS exchange works and Kdc test passed

+++ haox-asn1/src/main/java/org/haox/asn1/type/Asn1CollectionType.java
@@ -151,7 +151,7 @@ public abstract class Asn1CollectionType extends AbstractAsn1Type<Asn1Collection
     protected Integer getFieldAsInteger(int index) {
         Asn1Integer value = getFieldAs(index, Asn1Integer.class);
         if (value != null) {
-            return value.getValue().intValue();
+            return value.getValue();
         }
         return null;
     }
